Added Some VM tests + UI Search Bar Tests
Changes made based on PR Review:

helper functions in VM to change _uiState added
