+++ app/src/androidTest/java/com/github/se/eventradar/screens/HomeScreen.kt
@@ -1,6 +1,7 @@
 package com.github.se.eventradar.screens
 
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
+import androidx.compose.ui.test.hasTestTag
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.github.kakaocup.compose.node.element.KNode
 
@@ -16,4 +17,6 @@ class HomeScreen(semanticsProvider: SemanticsNodeInteractionsProvider) :
   val bottomNav: KNode = child { hasTestTag("bottomNavMenu") }
   val viewToggleFab: KNode = child { hasTestTag("viewToggleFab") }
   val map: KNode = child { hasTestTag("map") }
+  val searchBarAndFilter: KNode = child { hasTestTag("searchBarAndFilter") }
+  val filterPopUp: KNode = child { hasTestTag("filterPopUp") }
 }
